[ICU-6488] fix compiler warning X-Commit-URL: https://ssl.icu-project.org/trac/changeset/24454 Former-commit-id: ec11856f943ba82044c6ad524c77a9f6caf5d3d1

+++ icu4c/source/i18n/timezone.cpp
@@ -1430,7 +1430,7 @@ TimeZone::getTZDataVersion(UErrorCode& status)
             &len, &status);
 
         if (U_SUCCESS(status)) {
-            if (len >= sizeof(TZDATA_VERSION)) {
+            if (len >= (int32_t)sizeof(TZDATA_VERSION)) {
                 // Ensure that there is always space for a trailing nul in TZDATA_VERSION
                 len = sizeof(TZDATA_VERSION) - 1;
             }
